requirements.txt
# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data & Analytics
pandas>=2.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Database & Storage
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis>=5.0.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0

# Exchange Integration
python-binance>=1.0.0
ccxt>=4.0.0
websockets==12.0
aiohttp>=3.8.0

# Web Interface
jinja2==3.1.2
python-multipart==0.0.6

# Monitoring & Logging
structlog>=25.0.0
prometheus-client>=0.19.0

# Configuration
python-dotenv>=0.19.0
pyyaml>=6.0

# Note: ta-lib requires system-level installation: brew install ta-lib
# ta-lib>=0.4.0
